refactor: split moonutil common ownership modules
