+++ crates/moonutil/src/common.rs
@@ -19,7 +19,6 @@
 use crate::cond_expr::{CompileCondition, OptLevel};
 use crate::module::{MoonMod, MoonModJSON, MoonModRule};
 use crate::moon_pkg;
-use crate::mooncakes::ModuleName;
 use crate::package::{
     MoonPkg, MoonPkgJSON, SupportedTargetsDeclKind,
     convert_pkg_dsl_to_package_with_supported_targets_decl,
@@ -27,7 +26,6 @@ use crate::package::{
 };
 use anyhow::{Context, bail};
 use clap::ValueEnum;
-use fs4::fs_std::FileExt;
 use indexmap::IndexMap;
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
@@ -40,158 +38,12 @@ use std::io::{BufReader, BufWriter};
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
 
-pub const MOON_MOD: &str = "moon.mod";
-pub const MOON_MOD_JSON: &str = "moon.mod.json";
-pub const MOON_PKG_JSON: &str = "moon.pkg.json";
-pub const MOON_WORK: &str = "moon.work";
-pub const MOON_WORK_ENV: &str = "MOON_WORK";
-pub const MOON_NO_WORKSPACE: &str = "MOON_NO_WORKSPACE";
-pub const MOON_PKG: &str = "moon.pkg";
-pub const MBTI_GENERATED: &str = "pkg.generated.mbti";
-pub const MBTI_USER_WRITTEN: &str = "pkg.mbti";
-pub const MOONBITLANG_CORE: &str = "moonbitlang/core";
-pub const MOONBITLANG_CORE_BUILTIN: &str = "moonbitlang/core/builtin";
-pub const MOONBITLANG_CORE_PRELUDE: &str = "moonbitlang/core/prelude";
-pub const MOONBITLANG_COVERAGE: &str = "moonbitlang/core/coverage";
-pub const MOONBITLANG_ABORT: &str = "moonbitlang/core/abort";
-
-pub static MOD_NAME_STDLIB: ModuleName = ModuleName {
-    username: arcstr::literal!("moonbitlang"),
-    unqual: arcstr::literal!("core"),
-};
-
-pub const MOON_TEST_DELIMITER_BEGIN: &str = "----- BEGIN MOON TEST RESULT -----";
-pub const MOON_TEST_DELIMITER_END: &str = "----- END MOON TEST RESULT -----";
-
-pub const MOON_COVERAGE_DELIMITER_BEGIN: &str = "----- BEGIN MOONBIT COVERAGE -----";
-pub const MOON_COVERAGE_DELIMITER_END: &str = "----- END MOONBIT COVERAGE -----";
-
-pub const MOON_LOCK: &str = ".moon-lock";
-
-pub const WATCH_MODE_DIR: &str = "watch";
-
-pub const TEST_INFO_FILE: &str = "test_info.json";
-
-pub const WHITEBOX_TEST_PATCH: &str = "_wbtest.json";
-pub const BLACKBOX_TEST_PATCH: &str = "_test.json";
-
-pub const DOT_MBT_DOT_MD: &str = ".mbt.md";
-pub const DOT_MBTP: &str = ".mbtp";
-pub const DOT_MBL: &str = ".mbl";
-pub const DOT_MBY: &str = ".mby";
-
-pub const MOON_BIN_DIR: &str = "__moonbin__";
-
-pub const MOONCAKE_BIN: &str = "$mooncake_bin";
-pub const MOD_DIR: &str = "$mod_dir";
-pub const PKG_DIR: &str = "$pkg_dir";
-
-pub const SINGLE_FILE_TEST_PACKAGE: &str = "moon/test/single";
-pub const SINGLE_FILE_TEST_MODULE: &str = "moon/test";
-
-pub const SUB_PKG_POSTFIX: &str = "_sub";
-
-pub const PRELUDE_PROOF_DIR: &str = "prelude_proof";
-
-pub const O_EXT: &str = if cfg!(windows) { "obj" } else { "o" };
-#[allow(unused)]
-pub const DYN_EXT: &str = if cfg!(windows) {
-    "dll"
-} else if cfg!(target_os = "macos") {
-    "dylib"
-} else {
-    "so"
-};
-
-pub const A_EXT: &str = if cfg!(windows) { "lib" } else { "a" };
-
-pub fn is_moon_pkg_exist(dir: &Path) -> bool {
-    dir.join(MOON_PKG).exists() || dir.join(MOON_PKG_JSON).exists()
-}
-
-pub fn is_moon_pkg(filename: &str) -> bool {
-    filename == MOON_PKG || filename == MOON_PKG_JSON
-}
-
-pub fn is_moon_mod_exist(dir: &Path) -> bool {
-    dir.join(MOON_MOD).exists() || dir.join(MOON_MOD_JSON).exists()
-}
-
-pub fn is_moon_mod(filename: &str) -> bool {
-    filename == MOON_MOD || filename == MOON_MOD_JSON
-}
-
-pub fn is_moon_work(filename: &str) -> bool {
-    filename == MOON_WORK
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PackageSourceFileKind {
-    Mbt,
-    MbtMd,
-    Mbtp,
-    Mbl,
-    Mby,
-}
-
-pub fn package_source_file_kind(filename: &str) -> Option<PackageSourceFileKind> {
-    if filename.ends_with(".mbt") {
-        Some(PackageSourceFileKind::Mbt)
-    } else if filename.ends_with(DOT_MBT_DOT_MD) {
-        Some(PackageSourceFileKind::MbtMd)
-    } else if filename.ends_with(DOT_MBTP) {
-        Some(PackageSourceFileKind::Mbtp)
-    } else if filename.ends_with(DOT_MBL) {
-        Some(PackageSourceFileKind::Mbl)
-    } else if filename.ends_with(DOT_MBY) {
-        Some(PackageSourceFileKind::Mby)
-    } else {
-        None
-    }
-}
-
-pub fn is_watch_relevant_project_file(filename: &str) -> bool {
-    package_source_file_kind(filename).is_some()
-        || is_moon_pkg(filename)
-        || is_moon_mod(filename)
-        || is_moon_work(filename)
-}
-
-#[test]
-fn package_source_file_kind_detects_supported_package_inputs() {
-    assert_eq!(
-        package_source_file_kind("main.mbt"),
-        Some(PackageSourceFileKind::Mbt)
-    );
-    assert_eq!(
-        package_source_file_kind("guide.mbt.md"),
-        Some(PackageSourceFileKind::MbtMd)
-    );
-    assert_eq!(
-        package_source_file_kind("proof.mbtp"),
-        Some(PackageSourceFileKind::Mbtp)
-    );
-    assert_eq!(
-        package_source_file_kind("lexer.mbl"),
-        Some(PackageSourceFileKind::Mbl)
-    );
-    assert_eq!(
-        package_source_file_kind("parser.mby"),
-        Some(PackageSourceFileKind::Mby)
-    );
-    assert_eq!(package_source_file_kind("moon.pkg"), None);
-}
-
-#[test]
-fn watch_relevant_project_file_covers_sources_and_manifests() {
-    assert!(is_watch_relevant_project_file("moon.mod"));
-    assert!(is_watch_relevant_project_file("moon.mod.json"));
-    assert!(is_watch_relevant_project_file("moon.work"));
-    assert!(is_watch_relevant_project_file("moon.pkg"));
-    assert!(is_watch_relevant_project_file("moon.pkg.json"));
-    assert!(is_watch_relevant_project_file("lexer.mbl"));
-    assert!(!is_watch_relevant_project_file("README.md"));
-}
+pub use crate::constants::*;
+pub use crate::front_matter::*;
+pub use crate::glob::*;
+pub use crate::locks::FileLock;
+pub use crate::scripts::*;
+pub use crate::test_metadata::*;
 
 #[derive(Debug, Deserialize)]
 pub struct PatchJSON {
@@ -1382,158 +1234,6 @@ fn test_get_version() {
     assert!(!v.is_empty());
 }
 
-pub struct FileLock {
-    _file: std::fs::File,
-}
-
-impl Drop for FileLock {
-    fn drop(&mut self) {
-        fs4::fs_std::FileExt::unlock(&self._file).unwrap();
-    }
-}
-
-impl FileLock {
-    pub fn lock(path: &std::path::Path) -> std::io::Result<Self> {
-        Self::lock_with_verbosity(path, true)
-    }
-
-    pub fn lock_with_verbosity(path: &std::path::Path, verbose: bool) -> std::io::Result<Self> {
-        let file = std::fs::File::create(path.join(MOON_LOCK))?;
-        match file.try_lock_exclusive() {
-            Ok(_) => Ok(FileLock { _file: file }),
-            Err(_) => {
-                if verbose {
-                    #[cfg(not(test))]
-                    eprintln!(
-                        "Blocking waiting for file lock {} ...",
-                        path.join(MOON_LOCK).display()
-                    );
-                }
-                file.lock_exclusive()
-                    .map_err(|e| std::io::Error::new(e.kind(), "failed to lock target dir"))?;
-                Ok(FileLock { _file: file })
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum GeneratedTestDriver {
-    InternalTest(PathBuf),
-    WhiteboxTest(PathBuf),
-    BlackboxTest(PathBuf),
-}
-
-#[derive(Debug, ValueEnum, Clone, Copy)]
-pub enum DriverKind {
-    Internal,
-    Whitebox,
-    Blackbox,
-}
-
-impl std::fmt::Display for DriverKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let kind = match self {
-            Self::Internal => "internal",
-            Self::Whitebox => "whitebox",
-            Self::Blackbox => "blackbox",
-        };
-        write!(f, "{kind}")
-    }
-}
-
-#[derive(Debug, Default, ValueEnum, Clone, PartialEq, Copy, PartialOrd)]
-pub enum DiagnosticLevel {
-    Info,
-    #[value(alias = "warning")]
-    Warn,
-    #[default]
-    Error,
-}
-
-impl std::fmt::Display for DiagnosticLevel {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(match self {
-            Self::Error => "error",
-            Self::Warn => "warn",
-            Self::Info => "info",
-        })
-    }
-}
-
-pub const BLACKBOX_TEST_DRIVER: &str = "__generated_driver_for_blackbox_test.mbt";
-
-pub type FileName = String;
-pub type TestName = String;
-pub type TestBlockIndex = u32;
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct MbtTestInfo {
-    /// The index of the test block in the file, starting from 0.
-    pub index: TestBlockIndex,
-    /// The function name of the test block
-    pub func: String,
-    /// The name of the test block, if any
-    pub name: Option<TestName>,
-    /// The line number of the definition of the test block, if any
-    #[serde(default)]
-    pub line_number: Option<usize>,
-    /// The attributes of the test block (e.g., #cfg conditions)
-    #[serde(default)]
-    pub attrs: Vec<String>,
-}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct MooncGenTestInfo {
-    pub no_args_tests: IndexMap<FileName, Vec<MbtTestInfo>>,
-    pub with_args_tests: IndexMap<FileName, Vec<MbtTestInfo>>,
-    #[serde(default)] // for backward compatibility
-    pub with_bench_args_tests: IndexMap<FileName, Vec<MbtTestInfo>>,
-    #[serde(default)]
-    pub async_tests: IndexMap<FileName, Vec<MbtTestInfo>>,
-    #[serde(default)]
-    pub async_tests_with_args: IndexMap<FileName, Vec<MbtTestInfo>>,
-}
-
-impl MbtTestInfo {
-    pub fn has_skip(&self) -> bool {
-        self.attrs.iter().any(|attr| attr.starts_with("#skip"))
-    }
-}
-
-impl MooncGenTestInfo {
-    /// Convert part of the driver metadata into MoonBit declaraction code for
-    /// the test driver to use.
-    pub fn section_to_mbt(section: &IndexMap<FileName, Vec<MbtTestInfo>>) -> String {
-        use std::fmt::Write;
-
-        let mut result = String::new();
-        let default_name = "";
-
-        // Writing to string cannot fail, so unwrap() is safe here.
-        writeln!(result, "{{").unwrap();
-        for (file, tests) in section {
-            writeln!(result, "  \"{file}\": {{").unwrap();
-            for test in tests {
-                // tests with #skip attribute are also included in the driver, they will
-                // eventually be skipped by using cli arguments to the driver executable
-                writeln!(
-                    result,
-                    "    {}: ({}, [\"{}\"]),",
-                    test.index,
-                    test.func,
-                    test.name.as_deref().unwrap_or(default_name)
-                )
-                .unwrap();
-            }
-            writeln!(result, "  }},").unwrap();
-        }
-        writeln!(result, "}}").unwrap();
-
-        result
-    }
-}
-
 pub fn line_col_to_byte_idx(
     line_index: &line_index::LineIndex,
     line: u32,
@@ -1564,141 +1264,4 @@ impl StringExt for str {
     fn replace_crlf_to_lf(&self) -> String {
         self.replace("\r\n", "\n")
     }
-}
-
-pub enum PrePostBuild {
-    PreBuild,
-}
-
-impl PrePostBuild {
-    pub fn name(&self) -> String {
-        match self {
-            PrePostBuild::PreBuild => "pre-build".into(),
-        }
-    }
-
-    pub fn dbname(&self) -> String {
-        format!("{}.db", self.name())
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum IgnoredMoonScript {
-    Prebuild,
-    Postadd,
-}
-
-impl IgnoredMoonScript {
-    pub fn env_var(self) -> &'static str {
-        match self {
-            IgnoredMoonScript::Prebuild => "MOON_IGNORE_PREBUILD",
-            IgnoredMoonScript::Postadd => "MOON_IGNORE_POSTADD",
-        }
-    }
-}
-
-pub fn is_moon_script_ignored(script: IgnoredMoonScript) -> bool {
-    std::env::var_os(script.env_var()).is_some()
-}
-
-pub fn execute_postadd_script(dir: &Path) -> anyhow::Result<()> {
-    if is_moon_script_ignored(IgnoredMoonScript::Postadd) {
-        return Ok(());
-    }
-    let m = read_module_desc_file_in_dir(dir)?;
-    if let Some(scripts) = &m.scripts
-        && scripts.contains_key("postadd")
-    {
-        let postadd = scripts
-            .get("postadd")
-            .unwrap()
-            .split(' ')
-            .collect::<Vec<_>>();
-        if !postadd.is_empty() {
-            let command = postadd[0];
-            let args = &postadd[1..];
-            let output = std::process::Command::new(command)
-                .args(args)
-                .current_dir(dir)
-                .stdout(std::process::Stdio::inherit())
-                .stderr(std::process::Stdio::inherit())
-                .output()?;
-            if !output.status.success() {
-                bail!(
-                    "failed to execute postadd script in {},\ncommand: {},\n{}",
-                    dir.display(),
-                    command,
-                    String::from_utf8_lossy(&output.stderr)
-                );
-            }
-        }
-    }
-    Ok(())
-}
-
-#[derive(Debug, serde::Deserialize)]
-pub struct MbtMdHeader {
-    pub moonbit: Option<MbtMdSection>,
-}
-
-#[derive(Debug, serde::Deserialize, Default)]
-pub struct MbtMdSection {
-    pub deps: Option<IndexMap<String, crate::dependency::SourceDependencyInfo>>,
-    pub import: Option<crate::package::PkgJSONImport>,
-    pub backend: Option<String>,
-}
-
-pub fn parse_front_matter_config(single_file_path: &Path) -> anyhow::Result<Option<MbtMdHeader>> {
-    let single_file_string = single_file_path.display().to_string();
-    let front_matter_config: Option<MbtMdHeader> = if single_file_string.ends_with(DOT_MBT_DOT_MD) {
-        let content = std::fs::read_to_string(single_file_path)?;
-        let pattern = regex::Regex::new(r"(?s)^---\s*\n((?:[^\n]+\n)*?)---\s*\n")?;
-        if let Some(cap) = pattern.captures(&content) {
-            let yaml_content = cap.get(1).unwrap().as_str();
-            let config: MbtMdHeader = serde_yaml::from_str(yaml_content).map_err(|e| {
-                anyhow::anyhow!("Failed to parse front matter in markdown file: {}", e)
-            })?;
-
-            Some(config)
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-    Ok(front_matter_config)
-}
-
-/// Glob pattern matching supporting `*` (any sequence), `?` (any single character),
-/// and other glob patterns. Uses the `globset` crate for robust matching.
-pub enum GlobPatternMatcher<'a> {
-    Compiled(globset::GlobMatcher),
-    Literal(&'a str),
-}
-
-impl<'a> GlobPatternMatcher<'a> {
-    pub fn new(pattern: &'a str) -> Self {
-        use globset::GlobBuilder;
-        let glob = GlobBuilder::new(pattern)
-            .case_insensitive(false)
-            .literal_separator(false)
-            .build();
-        match glob {
-            Ok(glob) => Self::Compiled(glob.compile_matcher()),
-            // If pattern is invalid, fall back to literal comparison
-            Err(_) => Self::Literal(pattern),
-        }
-    }
-
-    pub fn is_match(&self, text: &str) -> bool {
-        match self {
-            Self::Compiled(matcher) => matcher.is_match(text),
-            Self::Literal(pattern) => pattern == &text,
-        }
-    }
-}
-
-/// Returns true if the text matches the pattern.
-pub fn glob_match(pattern: &str, text: &str) -> bool {
-    GlobPatternMatcher::new(pattern).is_match(text)
 }
