+++ crates/moonutil/src/lib.rs
@@ -24,14 +24,18 @@ pub mod cli;
 pub mod common;
 pub mod compiler_flags;
 pub mod cond_expr;
+pub mod constants;
 pub mod demangle;
 pub mod dependency;
 pub mod dirs;
 pub mod error_code_docs;
 pub mod features;
+pub mod front_matter;
 pub mod fuzzy_match;
 pub mod git;
+pub mod glob;
 pub mod graph;
+pub mod locks;
 pub mod module;
 pub mod moon_dir;
 pub mod moon_mod_patch;
@@ -41,8 +45,10 @@ pub mod package;
 pub mod path;
 pub mod platform;
 pub mod render;
+pub mod scripts;
 pub mod shlex;
 pub mod supported_targets;
+pub mod test_metadata;
 pub mod toolchain;
 pub mod user_warning;
 pub mod version;
